+++ db/src/postgres/migrations/2026-08-20-000000_task_resource_usage/down.sql
@@ -1,5 +1,5 @@
 ALTER TABLE tasks
-    DROP COLUMN peak_rss_bytes,
-    DROP COLUMN rss_total_bytes,
-    DROP COLUMN rss_sample_count,
+    DROP COLUMN peak_memory_bytes,
+    DROP COLUMN memory_total_bytes,
+    DROP COLUMN memory_sample_count,
     DROP COLUMN cpu_time_ms;